chore: rehydrate missing SQL from _quarantine; re-quarantine unrecoverable; manifest + stubs
