+++ public/usermood.php
@@ -1,82 +1,5 @@
 <?php
 declare(strict_types=1);
-require_once dirname(__DIR__) . '/bootstrap_web.php';
-
-use Pu239\Database;
-
-use Pu239\Cache;
-
-require_once __DIR__ . '/../include/bittorrent.php';
-$user = check_user_status();
-global $container;
-$db = $container->get(Database::class);, $site_config;
-$HTMLOUT = '';
-
-if (!isset($user['id'])) {
-    app_halt(_('log in to use this feature!'));
-}
-$more = $user['perms'] & UNLOCK_MORE_MOODS ? 2 : 1;
-if (isset($_GET['id'])) {
-    $moodid = (isset($_GET['id']) ? (int) $_GET['id'] : 1);
-    $res_moods = $db->run(');
-        write_log('<b>' . _('Mood Change') . '</b> ' . $user['username'] . ' ' . htmlsafechars($rmood['name']) . '<img src="' . $site_config['paths']['images_baseurl'] . 'smilies/' . htmlsafechars($rmood['image']) . '" alt="">');
-        $HTMLOUT = doc_head(_('moods')) . "
-        <link rel='stylesheet' href='" . get_file_name('vendor_css') . "'>
-        <link rel='stylesheet' href='" . get_file_name('css') . "'>
-        <link rel='stylesheet' href='" . get_file_name('main_css') . "'>
-      <script>
-      <!--
-      opener.location.reload(true);
-      self.close();
-      // -->
-      </script>";
-    } else {
-        app_halt(_('Hmmm. Invalid Mood choice.'));
-    }
-}
-$body_class = 'background-16 skin-2';
-$HTMLOUT .= doc_head(_('moods')) . "
-    <link rel='stylesheet' href='" . get_file_name('vendor_css') . "'>
-    <link rel='stylesheet' href='" . get_file_name('css') . "'>
-    <link rel='stylesheet' href='" . get_file_name('main_css') . "'>
-</head>";
-
-$body = '
-<body class="' . $body_class . '">
-    <script>
-        var theme = localStorage.getItem("theme");
-        if (theme) {
-            document.body.className = theme;
-        }
-    </script>';
-
-$div = '
-    <h3 class="has-text-centered has-text-primary top20">' . _fe("{0}'s Mood", $user['username']) . '</h3>
-    <div class="level-center bottom20">';
-$rows = $db->fetchAll('SELECT * FROM moods WHERE bonus < ' . sqlesc($more) . ' ORDER BY id');
-$count = 0;
-foreach ($rows as $arr) {
-    $div .= '
-        <span class="margin10 bordered has-text-centered bg-04">
-            <a href="?id=' . (int) $arr['id'] . '">
-                <img src="' . $site_config['paths']['images_baseurl'] . 'smilies/' . htmlsafechars($arr['image']) . '" alt="" class="bottom10">
-                <br>' . htmlsafechars($arr['name']) . '
-            </a>
-        </span>';
-}
-$div .= '
-    </div>';
-
-$body .= main_div($div) . '
-    <div class="w-100 has-text-centered margin20">
-        <a href="javascript:self.close();">
-            <span class="button bottom20">' . _('Close window') . '</span>
-        </a>
-    </div>
-    <noscript>
-        <a href="' . $site_config['paths']['baseurl'] . '/index.php">' . _('Back to site') . '</a>
-    </noscript>
-</body>';
-$HTMLOUT .= wrapper($body) . '
-</html>';
-echo $HTMLOUT;
+// TEMP STUB: original file re-quarantined due to missing SQL.
+// See tools/rehydrate_manifest.csv for details.
+throw new RuntimeException('Disabled: missing SQL; awaiting restoration.');
