+++ tools/rehydrate_log.txt
@@ -1,0 +1,3 @@
+Rehydration attempt summary:
+- 2025-09-20: Identified 16 corrupted scripts with truncated SQL ($db->run('); pattern).
+- No historical copies found under _quarantine; moved originals to _quarantine/rebroken/ and installed protective stubs.
